+++ RELEASE_NOTES.md
@@ -1,6 +1,6 @@
 ### May 2026
 
-- **Back Calculated Factor chart**: Updated by selected factor set.
+- **Back Calculated Factor charts**: Updated by selected factor set.
 - **Elapsed Time Comparison**: Added to handicap calculator.
 - **JSON View**: Re-added json view of items.
 - **PF Optimizer Invariants**: Ensure weighted ratios between spin, nonSpin and 2handed PF are preserved across optimisation